+++ crates/compiler/src/mml.rs
@@ -49,6 +49,7 @@ const _: () = assert!(
 const CHANNEL_NAMES: [&str; N_MUSIC_CHANNELS] = ["A", "B", "C", "D", "E", "F"];
 
 pub use self::bc_generator::MAX_BROKEN_CHORD_NOTES;
+pub use self::instruments::MmlInstrument;
 pub use self::tick_count_table::MmlTickCountTable;
 
 pub use self::metadata::MetaData;
@@ -210,6 +211,7 @@ pub fn compile_mml(
         song_data,
         duration,
         lines.sections,
+        instruments,
         channels,
         subroutines,
         #[cfg(feature = "mml_tracking")]