+++ crates/tad-gui/src/main.rs
@@ -59,7 +59,7 @@ use crate::tabs::{
     Tab, TabManager,
 };
 
-use audio_thread::{AudioMessage, AudioMonitor, ChannelsMask};
+use audio_thread::{AudioMessage, AudioMonitor, ChannelsMask, SongSkip};
 
 use compiler::data;
 use compiler::data::ProjectFile;
@@ -67,7 +67,6 @@ use compiler::driver_constants;
 use compiler::path::{ParentPathBuf, SourcePathBuf};
 use compiler::songs::SongData;
 use compiler::sound_effects::{convert_sfx_inputs_lossy, SoundEffectInput, SoundEffectsFile};
-use compiler::time::TickCounter;
 
 use compiler_thread::{PlaySampleArgs, SampleOutput};
 use files::{
@@ -153,7 +152,7 @@ pub enum GuiMessage {
     SongChanged(ItemId, String),
     RecompileSong(ItemId, String),
 
-    PlaySong(ItemId, String, Option<TickCounter>, ChannelsMask),
+    PlaySong(ItemId, String, Option<SongSkip>, ChannelsMask),
     PlaySoundEffect(ItemId),
     PlayInstrument(ItemId, PlaySampleArgs),
     PlaySample(ItemId, PlaySampleArgs),